Move snapshot source to extra/others/snapshot.
1. Update parameter buff to buf.
2. Add macro to disable lv_snapshot, enabled by default.
