+++ src/extra/others/lv_others.h
@@ -1,0 +1,38 @@
+/**
+ * @file lv_others.h
+ *
+ */
+
+#ifndef LV_OTHERS_H
+#define LV_OTHERS_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*********************
+ *      INCLUDES
+ *********************/
+#include "snapshot/lv_snapshot.h"
+
+/*********************
+ *      DEFINES
+ *********************/
+
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+/**********************
+ * GLOBAL PROTOTYPES
+ **********************/
+
+/**********************
+ *      MACROS
+ **********************/
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*LV_OTHERS_H*/
